locust/requirements.txt
bidict==0.22.1
Brotli==1.0.9
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
ConfigArgParse==1.5.3
Flask==2.2.2
Flask-BasicAuth==0.2.0
Flask-Cors==3.0.10
gevent==22.10.2
geventhttpclient==2.0.8
greenlet==2.0.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
locust==2.14.2
MarkupSafe==2.1.1
msgpack==1.0.4
psutil==5.9.4
pycparser==2.21
python-engineio==4.3.4
python-socketio==5.7.2
pywin32==305
pyzmq==25.0.0
requests==2.28.2
roundrobin==0.0.4
six==1.16.0
typing_extensions==4.4.0
urllib3==1.26.14
websocket-client==1.4.2
Werkzeug==2.2.2
zope.event==4.6
zope.interface==5.5.2
